third_party/WebKit/LayoutTests/pointer-lock/mouse-event-api.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
    var mouseEvent = document.createEvent("MouseEvent");
    shouldBeDefined("mouseEvent.movementX");
    shouldBeDefined("mouseEvent.movementY");

    // initializers are passed.
    shouldBe("new MouseEvent('eventType', { movementX: 888, movementY: 999 }).movementX", "888");
    shouldBe("new MouseEvent('eventType', { movementX: 888, movementY: 999 }).movementY", "999");
</script>
</body>
</html>
